Refactor
- add coverage requirements
- move msw setup before tests
- move api mock into __mocks__ folder
