+++ src/App.test.tsx
@@ -1,24 +1,9 @@
 import React from "react";
 import { renderConnected, within } from "../test/@testing-library/react";
-import { createMockServer } from "../test/msw/server";
 
 import App from "./App";
 
 const defaultProps = {};
-
-let server;
-beforeAll(() => {
-  server = createMockServer();
-  server.listen({
-    onUnhandledRequest(req) {
-      console.error(
-        "Found an unhandled %s request to %s",
-        req.method,
-        req.url.href
-      );
-    },
-  });
-});
 
 test("Loads the App to the Welcome screen", () => {
   const { getByTestId } = renderConnected(<App {...defaultProps} />);
