+++ test/jest.setup.ts
@@ -1,5 +1,6 @@
 import React from "react";
 import "@testing-library/jest-dom/extend-expect";
+import { createMockServer } from "./msw/server";
 
 global.React = React;
 
@@ -15,4 +16,15 @@ Object.defineProperty(window, "matchMedia", {
     removeEventListener: jest.fn(),
     dispatchEvent: jest.fn(),
   })),
+});
+
+const server = createMockServer();
+server.listen({
+  onUnhandledRequest(req) {
+    console.error(
+      "Found an unhandled %s request to %s",
+      req.method,
+      req.url.href
+    );
+  },
 });
