+++ test/msw/server.ts
@@ -1,10 +1,9 @@
-import { rest } from "msw";
 import { setupServer } from "msw/node";
 
-import commonHandlers from "./mockAPI";
+import commonHandlers from "../__mocks__/APIMock";
 
 function createMockServer(customHandlers = []) {
   return setupServer(...commonHandlers, ...customHandlers);
 }
 
-export { createMockServer, rest };
+export { createMockServer };
